feat: Validate DirectoryEntry
Added a `validate` endpoint which allows direct validation of post values before attempting to actually POST a directory entry. In future this could be used to do real time field uniqueness validation, but it is currently used to surface warnings about counts of consorial and managed institution entries in the database.

PR-1107